# specs/001-fullstack-web-todo/contracts/api.yaml
openapi: 3.0.3
info:
  title: Phase II Todo Application API
  description: |
    RESTful API for the Phase II Full-Stack Web Todo Application.
    Provides CRUD operations for tasks and categories with filtering, sorting, and search capabilities.
  version: 1.0.0
  contact:
    name: Hackathon II Team

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.todo.example.com
    description: Production server (future)

tags:
  - name: Tasks
    description: Task management operations
  - name: Categories
    description: Category management operations
  - name: Health
    description: API health checks

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  timestamp:
                    type: string
                    format: date-time

  /api/v1/tasks:
    get:
      tags: [Tasks]
      summary: List all tasks
      operationId: listTasks
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/Priority'
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          description: Search in title and description
          schema:
            type: string
        - name: due_before
          in: query
          schema:
            type: string
            format: date
        - name: due_after
          in: query
          schema:
            type: string
            format: date
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [title, priority, due_date, created_at]
            default: created_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'

    post:
      tags: [Tasks]
      summary: Create a new task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/ValidationError'

  /api/v1/tasks/{task_id}:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Tasks]
      summary: Get a task by ID
      operationId: getTask
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Tasks]
      summary: Update a task
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tasks]
      summary: Delete a task (soft delete)
      operationId: deleteTask
      responses:
        '204':
          description: Task deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/tasks/{task_id}/complete:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags: [Tasks]
      summary: Mark task as complete
      operationId: completeTask
      description: |
        Marks the task as completed. If the task has a recurrence pattern,
        a new task instance is automatically created for the next occurrence.
      responses:
        '200':
          description: Task completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
                  next_task:
                    $ref: '#/components/schemas/Task'
                    description: New recurring task instance (if applicable)
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/tasks/{task_id}/uncomplete:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags: [Tasks]
      summary: Mark task as pending
      operationId: uncompleteTask
      responses:
        '200':
          description: Task marked as pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/categories:
    get:
      tags: [Categories]
      summary: List all categories
      operationId: listCategories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'

    post:
      tags: [Categories]
      summary: Create a new category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/ValidationError'

  /api/v1/categories/{category_id}:
    parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Categories]
      summary: Get a category by ID
      operationId: getCategory
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Categories]
      summary: Update a category
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Categories]
      summary: Delete a category
      operationId: deleteCategory
      description: Cannot delete default (system) categories
      responses:
        '204':
          description: Category deleted successfully
        '400':
          description: Cannot delete default category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    TaskStatus:
      type: string
      enum: [pending, completed]
      default: pending

    Priority:
      type: string
      enum: [high, medium, low]
      default: medium

    RecurrenceType:
      type: string
      enum: [none, daily, weekly, monthly]
      default: none

    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        category_id:
          type: string
          format: uuid
          nullable: true
        category:
          $ref: '#/components/schemas/Category'
        due_date:
          type: string
          format: date
          nullable: true
        due_time:
          type: string
          format: time
          nullable: true
        recurrence_type:
          $ref: '#/components/schemas/RecurrenceType'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - title
        - status
        - priority
        - recurrence_type
        - created_at
        - updated_at

    TaskCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        priority:
          $ref: '#/components/schemas/Priority'
        category_id:
          type: string
          format: uuid
          nullable: true
        due_date:
          type: string
          format: date
          nullable: true
        due_time:
          type: string
          format: time
          nullable: true
        recurrence_type:
          $ref: '#/components/schemas/RecurrenceType'
      required:
        - title

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        priority:
          $ref: '#/components/schemas/Priority'
        category_id:
          type: string
          format: uuid
          nullable: true
        due_date:
          type: string
          format: date
          nullable: true
        due_time:
          type: string
          format: time
          nullable: true
        recurrence_type:
          $ref: '#/components/schemas/RecurrenceType'

    TaskResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Task'

    TaskListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 50
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          nullable: true
        is_default:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - is_default
        - created_at
        - updated_at

    CategoryCreate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          nullable: true
      required:
        - name

    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          nullable: true

    CategoryResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Category'

    CategoryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Category'

    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer

    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    ValidationError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Request validation failed
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found